build(): move event-data-loader to loaders directory
